Replicate all of the ViewModels methods into a static repository
